server/python_ai/requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv

langchain
langchain-community
langchain-groq
langchain-text-splitters

sentence-transformers
faiss-cpu
pymupdf

transformers
torch
